sib/pyproject.toml
[tool.poetry]
name = "sib"
version = "0.1.0"
description = ""
authors = ["vadikam100500 <vadik.msk.100500@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.2.9"
python-dotenv = "^0.19.2"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
django-extensions = "^3.1.5"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-broken-line = "^0.4.0"
flake8-return = "^1.1.3"
flake8-isort = "^4.1.1"
pep8-naming = "^0.12.1"
mypy = "^0.931"
isort = "^5.10.1"
pylint = "^2.12.2"
ipython = "7.31"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
